Modified int to long

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
 
-        int[] sale = {67, 87, 89};
+        long[] sale = {67, 87, 89};
         SalesManager manager = new SalesManager(sale);
         System.out.println(manager.max());
